Adds Particle System for Skills
Adds a particle system that can instantiate particles when skills are used (called by the skill's animation) and are deleted when the skill's animation ends. Adds the particle for Promise of Riches.Also adds the skill's art, animation and animator controller. Adds the art of instill fear, but its animation and controller are still missing.